requirements.txt
git+https://github.com/google/adk-python.git@main
fastapi
uvicorn
python-dotenv
